+++ diesel-migrations/2025-08-10-012300_add_backfill_to_brackets/up.sql
@@ -1,0 +1,2 @@
+-- Your SQL goes here
+ALTER TABLE brackets ADD COLUMN backfill_note TEXT NULL;