+++ src/schema.rs
@@ -1,6 +1,17 @@
 // @generated automatically by Diesel CLI.
 
 diesel::table! {
+    _sqlx_migrations (version) {
+        version -> Nullable<BigInt>,
+        description -> Text,
+        installed_on -> Timestamp,
+        success -> Bool,
+        checksum -> Binary,
+        execution_time -> BigInt,
+    }
+}
+
+diesel::table! {
     bracket_race_infos (id) {
         id -> Integer,
         bracket_race_id -> Integer,
@@ -44,6 +55,7 @@ diesel::table! {
         season_id -> Integer,
         state -> Text,
         bracket_type -> Text,
+        backfill_note -> Nullable<Text>,
     }
 }
 
@@ -159,6 +171,7 @@ diesel::joinable!(race_events -> bracket_race_infos (bracket_race_info_id));
 diesel::joinable!(race_runs -> races (race_id));
 
 diesel::allow_tables_to_appear_in_same_query!(
+    _sqlx_migrations,
     bracket_race_infos,
     bracket_races,
     bracket_rounds,